Clear existing pidfiles with --first-run
Pidfiles are not cleared on shutdown (and cannot be reliably cleared
ever due to power outakes etc. random things). By default Webmenu will
try to stop any existing Webmenu instances based on the pidfiles. After
reboot there will be a pidfile but no Webmenu instance which may cause
some random pid to be killed. So use --first-run to avoid this.